requirements.txt
aioredis==2.0.1
annotated-types==0.7.0
anyio==4.6.2
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
axios==0.4.0
beautifulsoup4==4.12.3
black==24.8.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==43.0.1
daphne==4.0.0
decorator==5.1.1
distro==1.9.0
dj-rest-auth==6.0.0
Django==4.2
django-allauth==65.0.2
django-cors-headers==4.4.0
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-redis==5.4.0
django-seed==0.3.1
django-storages==1.14.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
executing==2.1.0
Faker==30.0.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
hyperlink==21.0.0
iamport==0.0.1
idna==3.10
incremental==24.7.2
ipython==8.27.0
jedi==0.19.1
jiter==0.6.1
jmespath==1.0.1
lxml==5.3.0
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.1.0
mypy-extensions==1.0.0
oauthlib==3.2.2
openai==1.51.2
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.0
rsa==4.9
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.1
stack-data==0.6.3
tomli==2.0.1
toposort==1.10
tqdm==4.66.5
traitlets==5.14.3
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
zope.interface==7.1.0
